[수정]데이터 프롭스 전달 구조로 변경
