+++ views/screens/MyPage/LogList.js
@@ -53,45 +53,16 @@ const styles = StyleSheet.create({
   },
 });
 
-const LogList = list => {
+const LogList = ListData => {
   const [text, setText] = useState('Not Pressed');
-  const [listData, setListData] = useState(selectedData());
-  const LIST__DATA = Array();
 
-  function selectedData() {
-    if (list == '방문기록') return LIST_VIEW_DATA;
-    else if (list == '좋아요') return LIST_LIKE_DATA;
-    else if (list == '찜하기') return LIST_WANT_DATA;
-  }
-
-  var data = '2023.11.';
-  const LIST_VIEW_DATA = Array(30)
-    .fill('')
-    .map((_, i) => ({
-      key: `${i}`,
-      text: `김채린님이 회원님을 방문하였습니다.`,
-      date: `${data}${i + 1}`,
-    }));
-
-  const LIST_LIKE_DATA = Array(8)
-    .fill('')
-    .map((_, i) => ({
-      key: `${i}`,
-      text: `조호연님이 회원님의 리뷰에 좋아요를 눌렀습니다.`,
-    }));
-  const LIST_WANT_DATA = Array(20)
-    .fill('')
-    .map((_, i) => ({
-      key: `${i}`,
-      text: `김건우님이 회원님의 인터뷰 영상을 찜하였습니다.`,
-    }));
-
-  console.log(listData);
+  const result = ListData;
+  console.log(result);
 
   return (
     <SafeAreaView style={styles.container}>
       <SwipeListView
-        data={LIST_VIEW_DATA}
+        data={result}
         renderItem={({item}) => (
           <View style={styles.swipeListItem}>
             <Title color={'black'}>{item.text}</Title>
